+++ spec/macros4cuke/templating/placeholder_spec.rb
@@ -18,11 +18,11 @@ describe Placeholder do
   context 'Creation and initialization:' do
 
     it 'should be created with a variable name' do
-      ->() { Placeholder.new('foobar') }.should_not raise_error
+      expect { Placeholder.new('foobar') }.not_to raise_error
     end
 
     it 'should know the name of its variable' do
-      subject.name.should == 'foobar'
+      expect(subject.name).to eq('foobar')
     end
 
   end # context
@@ -31,11 +31,11 @@ describe Placeholder do
     it 'should render an empty string when no actual value is absent' do
       # Case: context has no value associated to 'foobar'
       rendered_text = subject.render(Object.new, {})
-      rendered_text.should be_empty
+      expect(rendered_text).to be_empty
 
       # Case: locals Hash has a nil value associated to 'foobar'
       rendered_text = subject.render(Object.new, { 'foobar' => nil })
-      rendered_text.should be_empty
+      expect(rendered_text).to be_empty
 
       # Case: context object has a nil value associated to 'foobar'
       context = Object.new
@@ -43,13 +43,13 @@ describe Placeholder do
         nil
       end
       rendered_text = subject.render(context, {})
-      rendered_text.should be_empty
+      expect(rendered_text).to be_empty
     end
 
     it 'should render the actual value bound to the placeholder' do
       # Case: locals Hash has a value associated to 'foobar'
       rendered_text = subject.render(Object.new, { 'foobar' => 'hello' })
-      rendered_text.should == 'hello'
+      expect(rendered_text).to eq('hello')
 
       # Case: context object has a value associated to 'foobar'
       context = Object.new
@@ -57,7 +57,7 @@ describe Placeholder do
         'world'
       end
       rendered_text = subject.render(context, {})
-      rendered_text.should == 'world'
+      expect(rendered_text).to eq('world')
     end
 
   end # context