+++ spec/spec_helper.rb
@@ -8,4 +8,12 @@ require 'simplecov'
 require 'rspec'	# Use the RSpec framework
 require 'pp'	# Use pretty-print for debugging purposes
 
+RSpec.configure do |config|
+  config.expect_with :rspec do |c|
+    # Disable the `should` syntax...
+    c.syntax = :expect
+  end
+end
+
+
 # End of file